+++ tests/integration_tests.rs
@@ -10,8 +10,8 @@ extern crate lazy_static;
 fn check_set_wifi_config() {
     let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
-    let config_file_path = tr.to_pathbuf("wpa_supplicant.conf");
-    let image_path = tr.to_pathbuf("image.wic");
+    let config_file_path = tr.to_pathbuf("testfiles/wpa_supplicant.conf");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
 
     assert_eq!(
         true,
@@ -24,8 +24,8 @@ fn check_set_wifi_config() {
 fn check_set_wifi_config_bmap() {
     let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
-    let config_file_path = tr.to_pathbuf("wpa_supplicant.conf");
-    let image_path = tr.to_pathbuf("image.wic");
+    let config_file_path = tr.to_pathbuf("testfiles/wpa_supplicant.conf");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
 
     assert_eq!(
         true,
@@ -37,8 +37,8 @@ fn check_set_wifi_config_bmap() {
 fn check_set_enrollment_config() {
     let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
-    let enrollment_config_file_path = tr.to_pathbuf("enrollment_static.json");
-    let image_path = tr.to_pathbuf("image.wic");
+    let enrollment_config_file_path = tr.to_pathbuf("testfiles/enrollment_static.json");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
 
     assert_eq!(
         true,
@@ -51,8 +51,8 @@ fn check_set_enrollment_config() {
 fn check_set_enrollment_config_bmap() {
     let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
-    let enrollment_config_file_path = tr.to_pathbuf("enrollment_static.json");
-    let image_path = tr.to_pathbuf("image.wic");
+    let enrollment_config_file_path = tr.to_pathbuf("testfiles/enrollment_static.json");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
 
     assert_eq!(
         true,
@@ -64,8 +64,8 @@ fn check_set_enrollment_config_bmap() {
 fn check_set_enrollment_config_missing_dps() {
     let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
-    let enrollment_config_file_path = tr.to_pathbuf("enrollment_static_missing_dps.json");
-    let image_path = tr.to_pathbuf("image.wic");
+    let enrollment_config_file_path = tr.to_pathbuf("testfiles/enrollment_static_missing_dps.json");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
 
     assert_ne!(
         None,
@@ -80,8 +80,9 @@ fn check_set_enrollment_config_missing_dps() {
 fn check_set_enrollment_config_missing_iothub() {
     let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
-    let enrollment_config_file_path = tr.to_pathbuf("enrollment_static_missing_iothub.json");
-    let image_path = tr.to_pathbuf("image.wic");
+    let enrollment_config_file_path =
+        tr.to_pathbuf("testfiles/enrollment_static_missing_iothub.json");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
 
     assert_ne!(
         None,
@@ -96,8 +97,8 @@ fn check_set_enrollment_config_missing_iothub() {
 fn check_set_enrollment_config_unknown_key() {
     let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
-    let enrollment_config_file_path = tr.to_pathbuf("enrollment_static_unknown_key.json");
-    let image_path = tr.to_pathbuf("image.wic");
+    let enrollment_config_file_path = tr.to_pathbuf("testfiles/enrollment_static_unknown_key.json");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
 
     assert_eq!(
         true,
@@ -110,8 +111,8 @@ fn check_set_enrollment_config_invalid_connection_string() {
     let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
     let enrollment_config_file_path =
-        tr.to_pathbuf("enrollment_static_invalid_connection_string.json");
-    let image_path = tr.to_pathbuf("image.wic");
+        tr.to_pathbuf("testfiles/enrollment_static_invalid_connection_string.json");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
 
     assert_ne!(
         None,
@@ -126,11 +127,11 @@ fn check_set_enrollment_config_invalid_connection_string() {
 fn check_set_identity_gateway_config() {
     let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
-    let config_file_path = tr.to_pathbuf("identity_config_gateway.toml");
-    let image_path = tr.to_pathbuf("image.wic");
-    let root_ca_file_path = tr.to_pathbuf("root.ca.cert.pem");
-    let edge_device_identity_full_chain_file_path = tr.to_pathbuf("full-chain.cert.pem");
-    let edge_device_identity_key_file_path = tr.to_pathbuf("device-ca.key.pem");
+    let config_file_path = tr.to_pathbuf("testfiles/identity_config_gateway.toml");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
+    let root_ca_file_path = tr.to_pathbuf("testfiles/root.ca.cert.pem");
+    let edge_device_identity_full_chain_file_path = tr.to_pathbuf("testfiles/full-chain.cert.pem");
+    let edge_device_identity_key_file_path = tr.to_pathbuf("testfiles/device-ca.key.pem");
 
     assert_eq!(
         true,
@@ -151,11 +152,11 @@ fn check_set_identity_gateway_config() {
 fn check_set_identity_gateway_config_bmap() {
     let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
-    let config_file_path = tr.to_pathbuf("identity_config_gateway.toml");
-    let image_path = tr.to_pathbuf("image.wic");
-    let root_ca_file_path = tr.to_pathbuf("root.ca.cert.pem");
-    let edge_device_identity_full_chain_file_path = tr.to_pathbuf("full-chain.cert.pem");
-    let edge_device_identity_key_file_path = tr.to_pathbuf("device-ca.key.pem");
+    let config_file_path = tr.to_pathbuf("testfiles/identity_config_gateway.toml");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
+    let root_ca_file_path = tr.to_pathbuf("testfiles/root.ca.cert.pem");
+    let edge_device_identity_full_chain_file_path = tr.to_pathbuf("testfiles/full-chain.cert.pem");
+    let edge_device_identity_key_file_path = tr.to_pathbuf("testfiles/device-ca.key.pem");
 
     assert_eq!(
         true,
@@ -175,9 +176,9 @@ fn check_set_identity_gateway_config_bmap() {
 fn check_set_identity_leaf_config() {
     let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
-    let config_file_path = tr.to_pathbuf("identity_config_leaf.toml");
-    let image_path = tr.to_pathbuf("image.wic");
-    let root_ca_file_path = tr.to_pathbuf("root.ca.cert.pem");
+    let config_file_path = tr.to_pathbuf("testfiles/identity_config_leaf.toml");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
+    let root_ca_file_path = tr.to_pathbuf("testfiles/root.ca.cert.pem");
 
     assert_eq!(
         true,
@@ -191,9 +192,9 @@ fn check_set_identity_leaf_config() {
 fn check_set_identity_leaf_config_bmap() {
     let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
-    let config_file_path = tr.to_pathbuf("identity_config_leaf.toml");
-    let image_path = tr.to_pathbuf("image.wic");
-    let root_ca_file_path = tr.to_pathbuf("root.ca.cert.pem");
+    let config_file_path = tr.to_pathbuf("testfiles/identity_config_leaf.toml");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
+    let root_ca_file_path = tr.to_pathbuf("testfiles/root.ca.cert.pem");
 
     assert_eq!(
         true,
@@ -206,8 +207,8 @@ fn check_set_identity_leaf_config_bmap() {
 fn check_set_identity_config() {
     let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
-    let config_file_path = tr.to_pathbuf("identity_config_dps_tpm.toml");
-    let image_path = tr.to_pathbuf("image.wic");
+    let config_file_path = tr.to_pathbuf("testfiles/identity_config_dps_tpm.toml");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
 
     assert_eq!(
         true,
@@ -220,8 +221,35 @@ fn check_set_identity_config() {
 fn check_set_identity_config_bmap() {
     let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
-    let config_file_path = tr.to_pathbuf("identity_config_dps_tpm.toml");
-    let image_path = tr.to_pathbuf("image.wic");
+    let config_file_path = tr.to_pathbuf("testfiles/identity_config_dps_tpm.toml");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
+
+    assert_eq!(
+        true,
+        docker::set_identity_config(&config_file_path, &image_path, true).is_ok()
+    );
+}
+
+#[test]
+fn check_set_identity_config_est_template() {
+    let tr = Testrunner::new(function_name!().split("::").last().unwrap());
+
+    let config_file_path = tr.to_pathbuf("conf/config.toml.est.template");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
+
+    assert_eq!(
+        true,
+        docker::set_identity_config(&config_file_path, &image_path, false).is_ok()
+    );
+}
+
+#[test]
+#[ignore]
+fn check_set_identity_config_est_template_bmap() {
+    let tr = Testrunner::new(function_name!().split("::").last().unwrap());
+
+    let config_file_path = tr.to_pathbuf("conf/config.toml.est.template");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
 
     assert_eq!(
         true,
@@ -233,8 +261,8 @@ fn check_set_identity_config_bmap() {
 fn check_set_device_cert() {
     let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
-    let intermediate_full_chain_crt_path = tr.to_pathbuf("test-int-ca_fullchain.pem");
-    let intermediate_full_chain_crt_key_path = tr.to_pathbuf("test-int-ca.key");
+    let intermediate_full_chain_crt_path = tr.to_pathbuf("testfiles/test-int-ca_fullchain.pem");
+    let intermediate_full_chain_crt_key_path = tr.to_pathbuf("testfiles/test-int-ca.key");
 
     let intermediate_full_chain_crt = std::fs::read_to_string(&intermediate_full_chain_crt_path)
         .expect("could not read intermediate full-chain-certificate");
@@ -252,7 +280,7 @@ fn check_set_device_cert() {
         .create_cert_and_key("bla", &None, 1)
         .expect("could not create new device certificate");
 
-    let image_path = tr.to_pathbuf("image.wic");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
 
     assert_eq!(
         true,
@@ -271,8 +299,8 @@ fn check_set_device_cert() {
 fn check_set_iot_hub_device_update_config() {
     let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
-    let adu_config_file_path = tr.to_pathbuf("du-config.json");
-    let image_path = tr.to_pathbuf("image.wic");
+    let adu_config_file_path = tr.to_pathbuf("testfiles/du-config.json");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
 
     assert_eq!(
         true,
@@ -285,8 +313,8 @@ fn check_set_iot_hub_device_update_config() {
 fn check_set_iot_hub_device_update_config_bmap() {
     let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
-    let adu_config_file_path = tr.to_pathbuf("du-config.json");
-    let image_path = tr.to_pathbuf("image.wic");
+    let adu_config_file_path = tr.to_pathbuf("testfiles/du-config.json");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
 
     assert_eq!(
         true,
@@ -298,8 +326,8 @@ fn check_set_iot_hub_device_update_config_bmap() {
 fn check_set_boot_config() {
     let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
-    let boot_config_file_path = tr.to_pathbuf("boot.scr");
-    let image_path = tr.to_pathbuf("image.wic");
+    let boot_config_file_path = tr.to_pathbuf("testfiles/boot.scr");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
 
     assert_eq!(
         true,
@@ -312,8 +340,8 @@ fn check_set_boot_config() {
 fn check_set_boot_config_bmap() {
     let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
-    let boot_config_file_path = tr.to_pathbuf("boot.scr");
-    let image_path = tr.to_pathbuf("image.wic");
+    let boot_config_file_path = tr.to_pathbuf("testfiles/boot.scr");
+    let image_path = tr.to_pathbuf("testfiles/image.wic");
 
     assert_eq!(
         true,